Update qem-event-build-functions.php
This change enable sthe display of deposit type. Currently although the deposit type can be set for each event it is not displayed when the event is viewed. Instead the default seting is always displayed.
The deposit type is available in the $custom variable but instead the $event variable is used. This change reads the 'deposittype' from the $custom variable and sets the display text accordingly.
Ideally it should be possible to configure the display text for each deposit type but this change does not provide for this.